+++ vulnerabilities/migrations/0016_update_cpe_url.py
@@ -1,5 +1,6 @@
 from django.db import migrations
 
+
 class Migration(migrations.Migration):
 
     def update_cpe_url(apps, schema_editor):