+++ vulnerabilities/migrations/0037_advisory_weaknesses_weakness.py
@@ -13,14 +13,17 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name='advisory',
             name='weaknesses',
-            field=models.JSONField(blank=True, default=list, help_text='A list of CWE ids'),
+            field=models.JSONField(
+                blank=True, default=list, help_text='A list of CWE ids'),
         ),
         migrations.CreateModel(
             name='Weakness',
             fields=[
-                ('id', models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
+                ('id', models.AutoField(auto_created=True,
+                 primary_key=True, serialize=False, verbose_name='ID')),
                 ('cwe_id', models.IntegerField(help_text='CWE id')),
-                ('vulnerabilities', models.ManyToManyField(related_name='weaknesses', to='vulnerabilities.vulnerability')),
+                ('vulnerabilities', models.ManyToManyField(
+                    related_name='weaknesses', to='vulnerabilities.vulnerability')),
             ],
         ),
     ]