+++ vulnerabilities/migrations/0045_remove_duplicated_purls_with_same_qualifiers.py
@@ -44,5 +44,6 @@ class Migration(migrations.Migration):
     ]
 
     operations = [
-        migrations.RunPython(remove_dupes, reverse_code=migrations.RunPython.noop),
+        migrations.RunPython(
+            remove_dupes, reverse_code=migrations.RunPython.noop),
     ]