+++ vulnerabilities/migrations/0049_alter_package_unique_together_and_more.py
@@ -13,7 +13,8 @@ class Migration(migrations.Migration):
         migrations.AlterUniqueTogether(
             name="package",
             unique_together={
-                ("type", "namespace", "name", "version", "subpath", "qualifiers_temp")
+                ("type", "namespace", "name", "version",
+                 "subpath", "qualifiers_temp")
             },
         ),
         migrations.AlterField(