+++ vulnerabilities/migrations/0050_copy_qualifiers_temp_back_to_qualifiers.py
@@ -16,20 +16,19 @@ class Migration(migrations.Migration):
             qualifiers_temp = package.qualifiers_temp
             package.qualifiers = qualifiers_temp
             updatables.append(package)
-        
+
         updated = Package.objects.bulk_update(
-            objs = updatables,
-            fields=["qualifiers",], 
+            objs=updatables,
+            fields=["qualifiers",],
             batch_size=500,
         )
-        print(f"Copied {updated} qualifiers_temp to qualifiers")            
-
-
+        print(f"Copied {updated} qualifiers_temp to qualifiers")
 
     dependencies = [
         ("vulnerabilities", "0049_alter_package_unique_together_and_more"),
     ]
 
     operations = [
-        migrations.RunPython(copy_qualifiers_temp, reverse_code=migrations.RunPython.noop),
+        migrations.RunPython(copy_qualifiers_temp,
+                             reverse_code=migrations.RunPython.noop),
     ]