+++ vulnerabilities/migrations/0052_remove_duped_purls_with_namespace_name_mixed.py
@@ -7,6 +7,7 @@ from django.db.models import Count
 
 from packageurl import PackageURL
 
+
 class Migration(migrations.Migration):
 
     def remove_dupes(apps, _):
@@ -51,7 +52,8 @@ class Migration(migrations.Migration):
     ]
 
     operations = [
-        migrations.RunPython(remove_dupes, reverse_code=migrations.RunPython.noop),
+        migrations.RunPython(
+            remove_dupes, reverse_code=migrations.RunPython.noop),
     ]