+++ CHANGELOG.rst
@@ -22,7 +22,7 @@ Version v34.0.0rc3
 Version v34.0.0rc2
 -------------------
 
-- We updated package-url models, WARNING: in next major version of 
+- We updated package-url models, WARNING: in next major version of
   vulnerablecode i.e v35.0.0 qualifiers will be of type ``string`` and not ``dict``.
 - We added changelog and dates on packages and vulnerabilities.
 - We fixed table borders in Vulnerability details UI #1356 (#1358)
@@ -38,7 +38,7 @@ Version v34.0.0rc2
 Version v34.0.0rc1
 -------------------
 
-- We updated package-url models, WARNING: in next major version of 
+- We updated package-url models, WARNING: in next major version of
   vulnerablecode i.e v35.0.0 qualifiers will be of type ``string`` and not ``dict``.
 - We added changelog and dates on packages and vulnerabilities.
 - We fixed table borders in Vulnerability details UI #1356 (#1358)